Qual: Fix phpstan notices (int expected)
# Qual: Fix phpstan notices (int expected)

Fix several notices where int was expected (casting/phpdoc/GETPOSTINT).
Separate phpstan exceptions for methods containing 'pdf' from
other methods/functions.

+++ scripts/emailings/mailing-send.php
@@ -6,6 +6,7 @@
  * Copyright (C) 2005-2016 Regis Houssin <regis.houssin@inodbox.com>
  * Copyright (C) 2019 		Nicolas ZABOURI	<info@inovea-conseil.com>
  * Copyright (C) 2024       Frédéric France             <frederic.france@free.fr>
+ * Copyright (C) 2024		MDW							<mdeweerd@users.noreply.github.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -396,7 +397,7 @@ if ($resql) {
 								}
 
 								if (getDolGlobalInt('MAILING_DELAY')) {
-									usleep((float) getDolGlobalInt('MAILING_DELAY') * 1000000);
+									usleep((int) ((float) getDolGlobalInt('MAILING_DELAY') * 1000000));
 								}
 							}
 						} else {
